+++ supvisors/tests/test_ttypes.py
@@ -32,8 +32,8 @@ def test_supvisors_instance_states():
 
 def test_supvisors_states():
     """ Test the SupvisorsStates enumeration. """
-    expected = ['OFF', 'INITIALIZATION', 'DEPLOYMENT', 'OPERATION', 'CONCILIATION', 'RESTARTING', 'RESTART',
-                'SHUTTING_DOWN', 'SHUTDOWN']
+    expected = ['OFF', 'INITIALIZATION', 'DEPLOYMENT', 'OPERATION', 'CONCILIATION', 'RESTARTING',
+                'SHUTTING_DOWN', 'FINAL']
     assert [x.name for x in SupvisorsStates] == expected